requirements.txt
# Core
python-dotenv==1.0.0
requests==2.31.0

# Data Processing
pandas==2.1.0

# Telegram Bot
python-telegram-bot==20.7

# Database
psycopg2-binary==2.9.9

# Logging
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
